Add policy tests
